backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
langgraph==0.2.*
langchain-core==0.3.*
langchain-google-genai==2.*
langchain-openai==0.3.*
google-generativeai==0.8.*
openai==1.*
httpx==0.28.*
pydantic==2.*
beautifulsoup4==4.*
python-dotenv==1.*
